+++ gardenPlannerBackend/gardenplanner/apps/garden/models.py
@@ -3,20 +3,6 @@ from django.contrib.auth.models import User
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-# Models will be implemented later
-# For example:
-# 
-# class Plant(models.Model):
-#     name = models.CharField(max_length=100)
-#     # other fields...
-# 
-# class Garden(models.Model):
-#     name = models.CharField(max_length=100)
-#     user = models.ForeignKey(User, on_delete=models.CASCADE)
-#     # other fields...
-#
-#
-# IMPORTANT NOTE: A Profile model must be implemented for storing location information. Please check RegisterSerializer in serializers.py
 
 class Profile(models.Model):
     ROLE_CHOICES = [
